Added a really crappy generator & hatches in types
I just wanted to get this running before sleep. I had to punch some
holes in the type system to get this to work as-is. Many things need
work in here:

I only have one type of node in the AST, the rest are nearly literal
translations of the tokens (sometimes literal).

Function calls are being configured like: `fn_FUNCTION_NAME_WITH_SUBS`
where I substitute `-` chars in the name with `_`. This will ultimately
cause name collisions so I need to be careful or reserve a char.

Generator tests are literal character matches and are not flexible
whatsoever.

There is no environment mapping here so there is no stdlib available for
the generator to use.

I am not handling mapping arithmetic and other operators over to JS
equivalents at this point (no env). This is something to handle

There is no scoping whatsoever.